+++ cmds/item.go
@@ -7,6 +7,7 @@ import (
 	"github.com/dottics/cli"
 	"github.com/google/uuid"
 	"log"
+	"time"
 )
 
 var ITEM *cli.Command
@@ -15,6 +16,7 @@ var itemsFilename = "items.json"
 type Item struct {
 	UUID            uuid.UUID `json:"uuid"`
 	TransactionUUID uuid.UUID `json:"transactionUUID"`
+	Date            time.Time `json:"date"`
 	Description     string    `json:"description"`
 	Debit           float64   `json:"debit"`
 	Credit          float64   `json:"credit"`
@@ -64,7 +66,7 @@ func (i *Item) RemoveTags(tags []string) {
 	i.Tags = keepTags
 }
 
-// TODO ITEMS: update the a map
+// TODO ITEMS: update to a map
 
 type Items []Item
 
@@ -86,13 +88,13 @@ func (xi *Items) Save() {
 
 // String returns a string representation of the items.
 func (xi *Items) String() string {
-	s := fmt.Sprintf("%-36s %-36s %-40s %-11s %-11s %s", "UUID", "TRANSACTION UUID", "DESCRIPTION", "DEBIT", "CREDIT", "TAGS")
+	s := fmt.Sprintf("%-36s %-36s %10s %-40s %-11s %-11s %s", "UUID", "TRANSACTION UUID", "DATE", "DESCRIPTION", "DEBIT", "CREDIT", "TAGS")
 	for _, i := range *xi {
 		desc := i.Description
 		if len(i.Description) > 40 {
 			desc = i.Description[:41]
 		}
-		s += fmt.Sprintf("%36s %36s %-40s %11.2f %11.2f %s\n", i.UUID, i.TransactionUUID, desc, i.Debit, i.Credit, i.Tags)
+		s += fmt.Sprintf("%36s %36s %10s %-40s %11.2f %11.2f %s\n", i.UUID, i.TransactionUUID, i.Date.Format("2006-01-02"), desc, i.Debit, i.Credit, i.Tags)
 	}
 	return s
 }