+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.7.2] - 2022-09-10
+### Changed
+- Updated items to now have the same date as the transaction, to simplify the 
+filtering of items.
 
 ## [0.7.1] - 2022-09-10
 ### Fixed